refactor(trello_03): fix adaptivity issues
