+++ src/scripts/components/clock.js
